Create README.md
Let's just practice with Git in this repository. Refer to the links under each block for more detailed walkthroughs.

First, you'll need to Clone it to your machine in order to get a local copy on your machine and also initiate Git's use of it for monitoring/version control. Click the green "Clone or Download" button above and then use the "Open in Desktop" option in order to clone the repo to your machine. This also tells the GitHub Desktop app to begin watching for changes to the repo.
https://help.github.com/desktop-beta/guides/contributing-to-projects/cloning-a-repository-from-github-to-github-desktop/

- Try adding text files or changing any existing ones that are in your local repository. Commit your changes to make a log entry of which changes you have made to your local files. Then Push the changes in the GitHubDesktop app for everyone else to see which changes you have made.
https://help.github.com/desktop-beta/guides/contributing-to-projects/committing-and-reviewing-changes-to-your-project/

- Try creating a branch and make separate changes from the Master branch. Later, the new branch will have to be remerged into the Master to make a common thread that everyone will be editing again. This is the beauty and the beast of version control.
https://help.github.com/desktop-beta/guides/contributing-to-projects/creating-a-branch-for-your-work/

- Try submitting a Pull Request to notify other collaborators in this repo that the changes you have made to the repo are ready to be pulled (downloaded by them and merged into their current version of the repo).
https://help.github.com/desktop-beta/guides/contributing-to-projects/creating-a-pull-request/